+++ src/Storage/Storage/File/Cmdlet/GetAzureStorageFile.cs
@@ -14,17 +14,15 @@
 
 namespace Microsoft.WindowsAzure.Commands.Storage.File.Cmdlet
 {
-    using Microsoft.Azure.Storage;
+    using global::Azure;
+    using global::Azure.Storage.Files.Shares;
+    using global::Azure.Storage.Files.Shares.Models;
     using Microsoft.Azure.Storage.File;
-    using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
     using Microsoft.WindowsAzure.Commands.Common.Storage.ResourceModel;
+    using System.Collections.Generic;
     using System.Globalization;
     using System.Management.Automation;
-    using System.Net;
 
-    [GenericBreakingChange("The cmdlet might run slower in a furture release since more file properties will be returned. Add parameter '-ExcludeExtendedInfo' to get same performance as before. " +
-        "The returned file properties will be moved from CloudFile to FileProperties (with '-Path'), ListFileProperties (without '-Path'). " +
-        "The returned Directory properties will be moved from CloudFileDirectory to ShareDirectoryProperties (with '-Path'), ListFileProperties (without '-Path').")]
     [Cmdlet("Get", Azure.Commands.ResourceManager.Common.AzureRMConstants.AzurePrefix + "StorageFile", DefaultParameterSetName = Constants.ShareNameParameterSetName)]
     [OutputType(typeof(AzureStorageFile))]
     public class GetAzureStorageFile : AzureStorageFileCmdletBase
@@ -64,6 +62,10 @@ namespace Microsoft.WindowsAzure.Commands.Storage.File.Cmdlet
             HelpMessage = "Path to an existing file/directory.")]
         public string Path { get; set; }
 
+
+        [Parameter(Mandatory = false, HelpMessage = "Not include extended file info like timestamps, ETag, attributes, permissionKey in list file and Directory.")]
+        public SwitchParameter ExcludeExtendedInfo { get; set; }
+
         public override void ExecuteCmdlet()
         {
             CloudFileDirectory baseDirectory;
@@ -85,64 +87,71 @@ namespace Microsoft.WindowsAzure.Commands.Storage.File.Cmdlet
                     throw new PSArgumentException(string.Format(CultureInfo.InvariantCulture, "Invalid parameter set name: {0}", this.ParameterSetName));
             }
 
+            // when only track1 object input, will miss storage context, so need to build storage context for prepare the output object.
+            if (this.Context == null)
+            {
+                this.Context = GetStorageContextFromTrack1FileServiceClient(baseDirectory.ServiceClient, DefaultContext);
+            }
+
+            ShareDirectoryClient baseDirClient = AzureStorageFileDirectory.GetTrack2FileDirClient(baseDirectory, ClientOptions);
+
             if (string.IsNullOrEmpty(this.Path))
             {
-                this.RunTask(async (taskId) =>
+                ShareDirectoryGetFilesAndDirectoriesOptions listFileOptions = new ShareDirectoryGetFilesAndDirectoriesOptions();
+                if (!this.ExcludeExtendedInfo.IsPresent)
                 {
-                    await this.Channel.EnumerateFilesAndDirectoriesAsync(
-                        baseDirectory,
-                        item => this.WriteListFileItemObject(taskId, this.Channel, item),
-                        this.RequestOptions,
-                        this.OperationContext,
-                        this.CmdletCancellationToken).ConfigureAwait(false);
-                });
+                    listFileOptions.Traits = ShareFileTraits.All;
+                    listFileOptions.IncludeExtendedInfo = true;
+                }
+                Pageable<ShareFileItem> fileItems = baseDirClient.GetFilesAndDirectories(listFileOptions, this.CmdletCancellationToken);
+                IEnumerable<Page<ShareFileItem>> fileitempages = fileItems.AsPages();
+                foreach (var page in fileitempages)
+                {
+                    foreach (var file in page.Values)
+                    {
+                        if (!file.IsDirectory) // is file
+                        {
+                            ShareFileClient shareFileClient = baseDirClient.GetFileClient(file.Name);
+                            WriteObject(new AzureStorageFile(shareFileClient, (AzureStorageContext)this.Context, file, ClientOptions)); 
+                        }
+                        else // Dir
+                        {
+                            ShareDirectoryClient shareDirClient = baseDirClient.GetSubdirectoryClient(file.Name);
+                            WriteObject(new AzureStorageFileDirectory(shareDirClient, (AzureStorageContext)this.Context, file, ClientOptions)); 
+                        }
+
+                    }
+                }
             }
             else
             {
-                this.RunTask(async (taskId) =>
+                if (ExcludeExtendedInfo.IsPresent)
                 {
-                    bool foundAFolder = true;
-                    string[] subfolders = NamingUtil.ValidatePath(this.Path);
-                    CloudFileDirectory targetDir = baseDirectory.GetDirectoryReferenceByPath(subfolders);
+                    WriteWarning("'-ExcludeExtendedInfo' will be omit, it only works when list file and directory without '-Path'.");
+                }
+                bool foundAFolder = true;
+                ShareDirectoryClient targetDir = baseDirClient.GetSubdirectoryClient(this.Path);
+                ShareDirectoryProperties dirProperties = null;
 
-                    try
-                    {
-                        await this.Channel.FetchDirectoryAttributesAsync(
-                            targetDir,
-                            null,
-                            this.RequestOptions,
-                            this.OperationContext,
-                            this.CmdletCancellationToken).ConfigureAwait(false);
-                    }
-                    catch (StorageException se)
-                    {
-                        if (null == se.RequestInformation
-                            || (int)HttpStatusCode.NotFound != se.RequestInformation.HttpStatusCode)
-                        {
-                            throw;
-                        }
+                try
+                {
+                    dirProperties = targetDir.GetProperties(this.CmdletCancellationToken).Value;
+                }
+                catch (global::Azure.RequestFailedException e) when (e.Status == 404)
+                {
+                    foundAFolder = false;
+                }
 
-                        foundAFolder = false;
-                    }
+                if (foundAFolder)
+                {
+                    WriteObject(new AzureStorageFileDirectory(targetDir, (AzureStorageContext)this.Context, dirProperties)); 
+                    return;
+                }
 
-                    if (foundAFolder)
-                    {
-                        WriteCloudFileDirectoryeObject(taskId, this.Channel, targetDir);
-                        return;
-                    }
+                ShareFileClient targetFile = baseDirClient.GetFileClient(this.Path);
+                ShareFileProperties fileProperties = targetFile.GetProperties(this.CmdletCancellationToken).Value;
 
-                    string[] filePath = NamingUtil.ValidatePath(this.Path, true);
-                    CloudFile targetFile = baseDirectory.GetFileReferenceByPath(filePath);
-
-                    await this.Channel.FetchFileAttributesAsync(
-                        targetFile,
-                        null,
-                        this.RequestOptions,
-                        this.OperationContext,
-                        this.CmdletCancellationToken).ConfigureAwait(false);
-
-                    WriteCloudFileObject(taskId, this.Channel, targetFile);
-                });
+                WriteObject(new AzureStorageFile(targetFile, (AzureStorageContext)this.Context, fileProperties, ClientOptions)); 
             }
         }
     }