+++ src/Storage/Storage.Management/ChangeLog.md
@@ -22,6 +22,8 @@
     -  `New-AzDataLakeGen2SasToken`
 * Show OAuth token in debug log in debug build only
     -  `New-AzStorageContext`
+* Supported return more file properties when list Azure file
+    -  `Get-AzStorageFile`
 
 ## Version 4.5.0
 * Supported DaysAfterLastTierChangeGreaterThan in Management Policy